Add :shared module
