+++ android-app/app/src/main/java/app/tivi/TiviApplication.kt
@@ -7,11 +7,13 @@ import android.app.Application
 import androidx.work.Configuration
 import androidx.work.WorkerFactory
 import app.tivi.extensions.unsafeLazy
-import app.tivi.inject.ApplicationComponent
+import app.tivi.inject.AndroidApplicationComponent
 import app.tivi.inject.create
 
 class TiviApplication : Application(), Configuration.Provider {
-    val component: ApplicationComponent by unsafeLazy { ApplicationComponent::class.create(this) }
+    val component: AndroidApplicationComponent by unsafeLazy {
+        AndroidApplicationComponent::class.create(this)
+    }
 
     private lateinit var workerFactory: WorkerFactory
