+++ android-app/app/src/main/java/app/tivi/inject/ActivityComponent.kt
@@ -9,6 +9,9 @@ import java.util.Locale
 import me.tatarka.inject.annotations.Provides
 
 interface ActivityComponent {
+    @get:Provides
+    val activity: Activity
+
     @Provides
     fun provideActivityLocale(activity: Activity): Locale {
         return ConfigurationCompat.getLocales(activity.resources.configuration)
